Added in h files
These files are used in the hypothesis chrome extension, and have
been added in here because they should be in the final produce.
For now, while I am softening out the edges and trying to make the
integration as smooth as possible, they are not relevant - I am adding
them here because it seems a good place to put them to remind myself
to add them in later.

The help/ folder is for when there is an issue with CORS - or, more
appropriately, when the user needs to enable extra priviledges for
h to run. This may be something we have to ask Beaglers to do.

The tests are for all of the files which are currently in lib, and
will need to be added in to the testing framework for Beagle and/
or added into their own modules with the lib files. This also goes
for karma.config.js.

The background.js and manifest.json files are the most relevant to
the current state of the Beagle chrome extension, and have been
added here as examples.