Move auth var, better require syntax, add domain key
This is related to trying to get nodemailer-mailgun-transport work better, which
for some unknown reason is bugging me today although it didn't yesterday. It's an
issue with brfs - I tried deleting shrinkwrap and starting again, but issues in
other submodules are popping up now. For instance, a path var in an fs.readFileSync
in regenerator that breaks brfs, and a document.body.append happening in
vm-browserify which unexpectedly happens before document exists. I'm not sure why
these issues are happening now, but I'm looking into them and hoping I don't have
to just start adding all of node_modules into git to solve these issues. #81