Added vm-browserify, pac-resolver, and regenerator patches
Regenerator had an issue with a variable being passed to fs.readFileSync.
This was fixed in .0.8.11, and then another bug was fixed in 0.8.13 that
was related to that patch. pac-resolver had an issue with using regenerator,
which was fixed in 1.2.5, and I'm currently waiting on a patch to up
regenerator version, which I'll then percolate up the chain back to
nodemailer-mailgun-transport. For now, npm-shrink everything, as there's
an open patch for vm-browserify to deal with dummy context not being
defined.