Store text selections with appropriate metadata
Before, each selection was stored dumbly as 'text'. Now, it uses a unique ID for the pdf, as well as the url it came from and an individual hash. The hash might be better done using the text itself put through a function, but that might end up with similarities, so... here we are.