Move the Milestones code
All of the React templates that were being required in main.js may have been
slowing the build down, and certianly weren't needed in the main chrome ext
as they are only used on the server. Also, the code in the HTML could be
browserified, too, so I set up a gulp task to do that and moved it all over.
Unfortunately, the code is still entirely borked, and I'm not sure why - the
functionality is exactly the same as it was in main.js, though, and at least
this gets this code out of the way for now.