Saved screenshot urls to localStorage #87
The Chrome API returns a one-off URL that I was hoping to be able to use in the
`inline` option for the Mailgun message. I'm not sure that this will ultimately
work, but for now I am saving the URL as part of the object with the text in
local Chrome storage. Having two databases is probably a bad idea, but I'm not
sure how else to do this intelligently. I see a few possibilities:
 - Litter the global scope with urls (and I don't know if they are persistent)
 - Save them to the database directly (again, persistence)
 - Save them to some sort of 'active' register in the database
 - Save them to local chrome storage as an active component
 - Save the images themselves and ditch the local URLs. Just use AWS for this.

@jbenet, do you have any ideas here?