Added in a location url
I often want to be able to link the URL to someone, but I have to exit out of the extension in order to find the original URL. This adds a small text link for the original url to the bottom of the sidebar if we're on a PDF. It's not a link because then it would open in the iFrame awkwardly. This closes #128.