Remove h folder
I had this in case I wanted to bring over the test files or the help page, but increasingly it is simply dead weight. At some point, we need to integrate in a testing platform, but it needn't be what they used, and I think the code has forked enough at this point that we probably should just start over from scratch. This code is still in the versioning history, so, no need to keep it around for now.