Better rendering added
This binds highlights to the window, and also renders on scroll. This fixes
some issues in https://github.com/BeagleLab/beagle-pdf/issues/4, but I
think it is probably still suboptimal in that it fetches the fingerprint on scroll
and doesn't remove highlights from the view, so they get checked every time.
This is probably slow.